[docker, npm] Add new standalone multi-stage docker build process
`build.Dockerfile` can build both the application and the runtime in one go. Mostly suitable for development. Can also be used in production to build a multi-platform image but extremely slow in non-native architectures.

Rebuilding is much faster due to the build caching using `cargo-chef`. Cache only invalidates if any of the manifest files are changed.